refactor(telemetry): satisfy Calor-first source policy
